extract UI from state, remove lingering global state references
