+++ src/system/state.rs
@@ -3,23 +3,15 @@
 //! Manages the robot's global state including:
 //! - Shared enums and cross-cutting state modules
 //!
-//! The state is protected by a mutex to ensure safe concurrent access
-//! from multiple tasks. All state changes are atomic and immediately
-//! visible to all tasks.
+//! State is compartmentalized into domain-specific modules, each with its own
+//! synchronization. Shared enums live here; domain state lives in dedicated
+//! submodules (e.g. `power`, `motion`, `perception`, `calibration`).
 //!
 //! # State Components
 //! - Operation Mode: Defined here and used by the `motion` module
 //! - Domain state lives in dedicated submodules
-//!
-//! # State Access Pattern
-//! ```rust
-//! let state = SYSTEM_STATE.lock().await;
-//! // Read or modify state here
-//! // Lock automatically released when state goes out of scope
-//! ```
 
 use defmt::Format;
-use embassy_sync::{blocking_mutex::raw::CriticalSectionRawMutex, mutex::Mutex};
 
 pub mod calibration;
 pub mod motion;
@@ -35,44 +27,6 @@ pub enum CalibrationStatus {
     Loaded,
     /// No calibration data exists in flash (needs calibration run)
     NotAvailable,
-}
-
-/// Top-level UI mode
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Format)]
-pub enum UiMode {
-    /// Main menu display
-    MainMenu,
-    /// System info screen (scrollable)
-    SystemInfo {
-        /// Current scroll offset into the system info line list
-        scroll_offset: u8,
-    },
-    /// Calibration submenu
-    CalibrateMenu,
-    /// Drive mode submenu
-    DriveModeMenu,
-    /// Test mode submenu
-    TestMenu,
-    /// Autonomous drive mode is active
-    RunningAutonomous {
-        /// Which drive mode is running
-        mode: DriveMode,
-    },
-    /// Combined test sequence running
-    RunningTest,
-    /// IMU test mode (live display)
-    RunningImuTest,
-    /// Basic motor test mode
-    RunningBasicMotorTest,
-    /// IR + ultrasonic live test mode
-    RunningIrUltrasonicTest,
-    /// Ultrasonic sweep test mode
-    RunningUltrasonicSweepTest,
-    /// Optional: calibration running state
-    Calibrating {
-        /// Selected calibration kind
-        kind: CalibrationSelection,
-    },
 }
 
 /// Main menu selections
@@ -122,19 +76,6 @@ pub enum CalibrationSelection {
     /// Gyroscope calibration
     Gyro,
 }
-
-/// Global system state protected by a mutex
-///
-/// Initialized to an empty struct; domain state lives in submodules.
-pub static SYSTEM_STATE: Mutex<CriticalSectionRawMutex, SystemState> = Mutex::new(SystemState {});
-
-/// Robot system state containing all runtime state information
-///
-/// This struct represents the complete state of the robot at any given time.
-/// Changes to these values trigger corresponding system behaviors through
-/// the event system.
-#[derive(Format)]
-pub struct SystemState {}
 
 // Motion state helpers moved to the `motion` module.
