+++ src/task/ui/state.rs
@@ -2,7 +2,47 @@
 
 use embassy_sync::{blocking_mutex::raw::CriticalSectionRawMutex, mutex::Mutex};
 
-use crate::system::state::UiMode;
+use defmt::Format;
+
+use crate::system::state::{CalibrationSelection, DriveMode};
+
+/// Top-level UI mode
+#[derive(Debug, Clone, Copy, Eq, PartialEq, Format)]
+pub enum UiMode {
+    /// Main menu display
+    MainMenu,
+    /// System info screen (scrollable)
+    SystemInfo {
+        /// Current scroll offset into the system info line list
+        scroll_offset: u8,
+    },
+    /// Calibration submenu
+    CalibrateMenu,
+    /// Drive mode submenu
+    DriveModeMenu,
+    /// Test mode submenu
+    TestMenu,
+    /// Autonomous drive mode is active
+    RunningAutonomous {
+        /// Which drive mode is running
+        mode: DriveMode,
+    },
+    /// Combined test sequence running
+    RunningTest,
+    /// IMU test mode (live display)
+    RunningImuTest,
+    /// Basic motor test mode
+    RunningBasicMotorTest,
+    /// IR + ultrasonic live test mode
+    RunningIrUltrasonicTest,
+    /// Ultrasonic sweep test mode
+    RunningUltrasonicSweepTest,
+    /// Optional: calibration running state
+    Calibrating {
+        /// Selected calibration kind
+        kind: CalibrationSelection,
+    },
+}
 
 /// UI state owned by the UI controller.
 #[derive(Clone, Copy)]
